requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
discord.py-self==1.9.2
idna==3.4
imgurpython==1.1.7
multidict==6.0.2
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
Telethon==1.26.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.1
